* commit '526fe8d45d36f5a8e33aedf14640e3d8ef58b28e':
  Size and offset fixes to compile for 64b host

+++ runtime/parsed_options.cc
@@ -288,13 +288,13 @@ bool ParsedOptions::Parse(const Runtime::Options& options, bool ignore_unrecogni
         return false;
       }
     } else if (StartsWith(option, "-XX:LongPauseLogThreshold=")) {
-      size_t value;
+      unsigned int value;
       if (!ParseUnsignedInteger(option, '=', &value)) {
         return false;
       }
       long_pause_log_threshold_ = MsToNs(value);
     } else if (StartsWith(option, "-XX:LongGCLogThreshold=")) {
-      size_t value;
+      unsigned int value;
       if (!ParseUnsignedInteger(option, '=', &value)) {
         return false;
       }
